Implementacion de los filtros por medio de un modelo de clase en vez de variables unitarias en los MVVM. Cambios en los servicios (Servicio Control de Usuario) para la adaptacion de los filtros. Nuevo metodo en la clase usuario para la obtencio generica de la Persona asociada a la clase. La funcion mostrarNotificacion del AbstractViewModel acepta como parametros al componente de referencia donde se posicionara. Mejoras y validaciones en el formulario para crear nuevos usuarios con los filtros de la grid incluidos. Los parametros "total" de los servicios se usaran en Integer a partir de esta fecha. Creacion del formulario para editar al usuario respectivo, se incluyeron las validaciones del formulario.

+++ proyectos/ProyectoPrueba/src/main/java/com/prueba/web/mvvm/controladores/seguridad/configuracion/ListaGruposViewModel.java
@@ -80,12 +80,12 @@ public class ListaGruposViewModel extends AbstractViewModel implements EventList
 			@BindingParam("sortDirection") Boolean sortDirection){
 		
 		Map<String, Object> parametros = servicioControlGrupo.consultarGrupos(grupoFiltro, fieldSort, sortDirection, page, PAGE_SIZE);
-		Long total = (Long) parametros.get("total");
+		Integer total =  (Integer) parametros.get("total");
 		grupos = (List<Group>) parametros.get("grupos");
 		gridGrupos.setMultiple(true);
 		gridGrupos.setCheckmark(true);
 		pagGrupos.setActivePage(page);
-		pagGrupos.setTotalSize(total.intValue());
+		pagGrupos.setTotalSize(total);
 	}
 	
 	/**COMMAND*/
